Added a (temporary) recipe and loot table entry for the sonic security system. Also removed an unnecessary cast and added the placement sound

+++ src/main/java/net/geforcemods/securitycraft/renderers/SonicSecuritySystemTileEntityRenderer.java
@@ -53,7 +53,7 @@ public class SonicSecuritySystemTileEntityRenderer extends TileEntityRenderer<So
 		}
 
 		matrix.rotate(POSITIVE_X_180);
-		modelSonicSecuritySystem.radar.rotateAngleY = (float)te.radarRotationDegrees;
+		modelSonicSecuritySystem.radar.rotateAngleY = te.radarRotationDegrees;
 		modelSonicSecuritySystem.render(matrix, buffer.getBuffer(RenderType.getEntitySolid(sonicSecuritySystemTexture)), p_225616_5_, OverlayTexture.NO_OVERLAY, 1.0F, 1.0F, 1.0F, 1.0F);
 	}
 }
